waste in objects.py
